+++ src/Ext/WarheadType/Body.h
@@ -49,16 +49,16 @@ public:
 		Valueable<bool> PenetratesIronCurtain;
 		Nullable<bool> PenetratesForceShield;
 
-		Valueable<double> Crit_Chance;
+		ValueableVector<double> Crit_Chance;
 		Valueable<bool> Crit_ApplyChancePerTarget;
-		Valueable<int> Crit_ExtraDamage;
-		Valueable<WarheadTypeClass*> Crit_Warhead;
+		ValueableVector<int> Crit_ExtraDamage;
+		Nullable<WarheadTypeClass*> Crit_Warhead;
 		Valueable<AffectedTarget> Crit_Affects;
 		Valueable<AffectedHouse> Crit_AffectsHouses;
 		ValueableVector<AnimTypeClass*> Crit_AnimList;
 		Nullable<bool> Crit_AnimList_PickRandom;
 		Valueable<bool> Crit_AnimOnAffectedTargets;
-		Valueable<double> Crit_AffectBelowPercent;
+		ValueableVector<double> Crit_AffectBelowPercent;
 		Valueable<bool> Crit_SuppressWhenIntercepted;
 
 		Nullable<AnimTypeClass*> MindControl_Anim;
@@ -146,7 +146,8 @@ public:
 		Valueable<bool> EffectsRequireVerses;
 
 		double Crit_RandomBuffer;
-		double Crit_CurrentChance;
+		std::vector<double> Crit_CurrentChance;
+		std::vector<int> Crit_CurrentExtraDamage;
 		bool Crit_Active;
 		bool WasDetonatedOnAllMapObjects;
 		bool Splashed;
@@ -188,16 +189,16 @@ public:
 			, PenetratesIronCurtain { false }
 			, PenetratesForceShield {}
 
-			, Crit_Chance { 0.0 }
+			, Crit_Chance {}
 			, Crit_ApplyChancePerTarget { false }
-			, Crit_ExtraDamage { 0 }
+			, Crit_ExtraDamage {}
 			, Crit_Warhead {}
 			, Crit_Affects { AffectedTarget::All }
 			, Crit_AffectsHouses { AffectedHouse::All }
 			, Crit_AnimList {}
 			, Crit_AnimList_PickRandom {}
 			, Crit_AnimOnAffectedTargets { false }
-			, Crit_AffectBelowPercent { 1.0 }
+			, Crit_AffectBelowPercent {}
 			, Crit_SuppressWhenIntercepted { false }
 
 			, MindControl_Anim {}
@@ -283,7 +284,8 @@ public:
 			, EffectsRequireVerses { true }
 
 			, Crit_RandomBuffer { 0.0 }
-			, Crit_CurrentChance { 0.0 }
+			, Crit_CurrentChance {}
+			, Crit_CurrentExtraDamage {}
 			, Crit_Active { false }
 			, WasDetonatedOnAllMapObjects { false }
 			, Splashed { false }
@@ -322,7 +324,7 @@ public:
 		void ApplyCrit(HouseClass* pHouse, TechnoClass* pTarget, TechnoClass* Owner, TechnoExt::ExtData* pTargetExt);
 		void ApplyShieldModifiers(TechnoClass* pTarget, TechnoExt::ExtData* pTargetExt);
 		void ApplyAttachEffects(TechnoClass* pTarget, HouseClass* pInvokerHouse, TechnoClass* pInvoker);
-		double GetCritChance(TechnoClass* pFirer) const;
+		std::pair<std::vector<double>, std::vector<int>> GetCritChanceAndExtraDamage(TechnoClass* pFirer) const;
 	};
 
 	class ExtContainer final : public Container<WarheadTypeExt>