+++ docs/Whats-New.md
@@ -437,6 +437,8 @@ New:
 - Option for Warhead damage to penetrate Iron Curtain or Force Shield (by Starkku)
 - Option for Warhead to remove all shield types at once (by Starkku)
 - Allow customizing voxel light source position (by Kerbiter, Morton, based on knowledge of thomassnedon)
+- Crit chance and extra damage based on target health percentage (by Ollerus)
+- Crit extra damage multipliers for new AttachEffect (by Ollerus)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)